+++ src/filecont.h
@@ -46,8 +46,8 @@ public:
     virtual bool  FindFile(const char *fileName, FlexDirEntry &entry) = 0;
     virtual bool  DeleteFile(const char *fileName) = 0;
     virtual bool  RenameFile(const char *oldName, const char *newName) = 0;
-    virtual bool  SetAttributes(const char *fileName, int setMask = 0,
-                                int clearMask = ~0) = 0;
+    virtual bool  SetAttributes(const char *fileName, Byte setMask = 0,
+                                Byte clearMask = ~0) = 0;
     virtual FlexFileBuffer ReadToBuffer(const char *fileName) = 0;
     virtual bool  WriteFromBuffer(const FlexFileBuffer &buffer,
                                   const char *fileName = nullptr) = 0;
