+++ src/idircnt.h
@@ -50,7 +50,7 @@ private:
     bool DeleteCurrent();
     bool RenameCurrent(const char *);
     bool SetDateCurrent(const BDate &date);
-    bool SetAttributesCurrent(int attributes);
+    bool SetAttributesCurrent(Byte attributes);
 
     DirectoryContainerIteratorImp();
     DirectoryContainerIteratorImp(const DirectoryContainerIteratorImp &);
